[VM] Expose WriteOpConverter
