+++ aptos-move/aptos-vm/src/block_executor/vm_wrapper.rs
@@ -7,13 +7,16 @@ use crate::{
     aptos_vm::AptosVM,
     block_executor::AptosTransactionOutput,
     data_cache::StorageAdapter,
+    move_vm_ext::write_op_converter::WriteOpConverter,
 };
 use aptos_block_executor::task::{ExecutionStatus, ExecutorTask};
 use aptos_logger::{enabled, Level};
 use aptos_mvhashmap::types::TxnIndex;
 use aptos_state_view::StateView;
+use aptos_types::{state_store::state_key::StateKey, write_set::WriteOp};
 use aptos_vm_logging::{log_schema::AdapterLogSchema, prelude::*};
 use move_core_types::{
+    effects::Op as MoveStorageOp,
     ident_str,
     language_storage::{ModuleId, CORE_CODE_ADDRESS},
     vm_status::VMStatus,
@@ -110,5 +113,32 @@ impl<'a, S: 'a + StateView + Sync> ExecutorTask for AptosExecutorTask<'a, S> {
             },
             Err(err) => ExecutionStatus::Abort(err),
         }
+    }
+
+    fn convert_to_value(
+        &self,
+        view: &impl StateView,
+        key: &StateKey,
+        maybe_blob: Option<Vec<u8>>,
+        creation: bool,
+    ) -> anyhow::Result<WriteOp> {
+        let storage_adapter = self.vm.as_move_resolver(view);
+        let wop_converter =
+            WriteOpConverter::new(&storage_adapter, self.vm.is_storage_slot_metadata_enabled());
+
+        let move_op = match maybe_blob {
+            Some(blob) => {
+                if creation {
+                    MoveStorageOp::New(blob)
+                } else {
+                    MoveStorageOp::Modify(blob)
+                }
+            },
+            None => MoveStorageOp::Delete,
+        };
+
+        wop_converter
+            .convert(key, move_op, false)
+            .map_err(|_| anyhow::Error::msg("Error on converting to WriteOp"))
     }
 }
